facerecog_tensorflow_gcp/gcf_face_comparison/requirements.txt
numpy
google-cloud-bigquery
google-api-python-client
oauth2client

